Report error if image size width fails
On non-Apple platforms, we don't currently have a way of determining the
width and height of an image, as we have not yet integrated a suitable
library. Add a parameter to DFPlatformGetImageDimensions which, where
this is not implemented, reports an error indicating the function is not
supported on the current platform.