Move string code from DFFilesytem.c to DFString.c
Move the following functions, which have nothing to do with filesystem
access, into DFString.c

DFFormatDouble
DFFormatDoublePct
DFFormatDoublePt
DFNormalizeWhitespace (now DFStringNormalizeWhitespace)

Background: DFFilesystem.c used to be called DFUtil.c and was a
catch-all places for functions that didn't fit anywhere else. When
renaming what used to be called "convertdoc" to "dfutil", the file was
renamed to avoid confusion between the two. The above functions were
still left in the file, but really belong in DFString.c.