Fix incorrect use of va_end
There were two cases in which we were calling va_end with a va_list
variable that was an original, not a copy. When two such calls were made
in succession on Windows, the second would fail as the pointer would no
longer be valid. This was not happening under OS X or Linux due to the
different way in which these platforms implement the stdarg macros.