Win32: Work around const typecheck bug in VC++
Visual C++'s type checker doesn't fully understand what the 'const'
keyword means in the context of pointers to pointers, which results in
false warnings when calling free.

In many places in the code, we use arrays of constant strings, e.g.:

    const char **keys = DFHashTableCopyKeys(x);

Here, the keys array itself is modifiable (we can assign to keys[n]),
but the strings *in* the array are constant (we can't assign to
keys[n][m]). After doing something with the array (which sometimes
includes modifying it; in many cases we sort the list of keys), we then
do the following:

    free(keys);

This is fine, because we're freeing non-constant memory. But when
typechecking the call, Visual C++ sees the const and incorrectly thinks
it applies to the array, rather than its contents, and issues a warning.

To avoid these spurious warnings, set a pragma to disable warning 4090
('function': different 'const' qualifiers).