Win32: Avoid unicode escapes in string literals
In GCC and clang we can write escape sequences of the form \uXXXX in
strings to express unicode characters. However Visual C++ doesn't like
this - or, at least, it complains about the characters not being
supported in the current codepage.

There's only two instances in which we actually need to do this -
specifying the square and bullet characters for lists in Word. Just
store the UTF-8 encoded characters in an array, and use that instead.