Win32: Use rand instead of random
The OS X manpages describe the rand function as a "bad random number
generator" and the random function as a "better random number
generator". The latter is not available on windows however. Of course
there are more secure alternatives available, but for our purposes we
don't need security.

The one place where this is used in the code is to generate a unique
prefix for the id attributes in HTML files used for mapping between
abstract and concrete elements when doing bidirectional transformation.

The motivation for having different prefixes, in the context of UX
Write, is that if someone copies some text from one word document
(converted to HTML) and then pastes it into another document (also
converted to HTML), then when the latter is saved, we don't want the id
attributes in the pasted elements to be confused for elements that
correspond to existing parts of the second word document. We don't need
high-quality random number generation for this.