Word: Merge WordPackage{AddMissingDocParts,SaveTo}
WordPackageAddMissingDocParts used to be responsible for ensuring we had
OPC parts and relationships for any DFDocument object in the
WordPackage. Because this function was only ever called from
WordPackageSaveTo, and we had duplicate NULL checks for the documents
(which are about to get more complicated), this logic has now been
combined with that of actually saving the relevant parts.