XML: Avoid incorrect text node indentation
When all an element's children are text nodes, seralise them without
indentation.

This affects documents containing emoji characters, where libxml2 seems
to parse these as separate text nodes. If you open a document and then
save it, the result will be written out with indentation added
unnecessarily between the nodes. This does not affect any of the tests
with the current implementation, but when we switch to saving + loading
the XML files before converting to plain text, it affects the test
basic/word/get-emoji02.test.