Word: Don't set update fields flag in fromPlain
When constructing a WordPackage object from a text package, avoid
setting the update fields flag in settings.xml. This was only relevant
when doing manual conversion from the command line and then opening the
resulting .docx file in Word. However it causes problems with the tests
when, with code soon to be committed, we save the package and then
construct a text package from the filesystem, since it results in
spurious settings.xml inclusions where we don't want them.