Word: Remove OPCPart fields from WordPackage
With the exception of the document part (to which all others are
related), don't store pointers to the OPCPart objects for the document
in the WordPackage object. They're not needed for the core part of the
conversion, and their presence there introduces unnecessary complexity
in the form of requiring them to be set at the right time during various
operations.

Instead, we now look up the parts as necessary. This happens in three
places:

1. WordPackageOpenFrom, where we need the part to find the URL from
   which to read the data of the XML file

2. WordPackageSaveTo, where we add any missing parts for XML documents
   that have been created during conversion

3. Word_fromPackage (called from Word_fromPlain), where we set the
   overrides in [Content_Types].xml directly