Word: Retain margin-left for List Paragraph style
By default, when you create a list in Word, it assigns all paragraphs
the "List Paragraph" style, which has a default indentation of 36
points. We exclude this from the generated HTML code, since lists in
HTML are automatically indented by default. However, the removal of this
property for the list paragraph style meant that on update, the word
document would no longer have it set.

With this commit, if we see margin-left set on the style, we save the
old copy under the name -word-margin-left, and clear margin-left. Then
on update, we inspect the value of -word-margin-left, and if it is set,
and margin-left is missing, we use -word-margin-left instead.

This fixes null upates in some cases, e.g. in
numbering/word/nullupdate-list01.test