Word: fromPlain: Work from zip file
Step 1 on the path to not using any of the Word APIs directly in dfutil.
Here, we first create a WordPackage as normal, directly accessing the
Word/OPC APIs and data structures in Word_fromPackage. We then save it
to a .docx (zip) file, and then open a new word package from that file.
This ensures that all the OPC logic now becomes part of the code path
run during tests.

The next step will to be directly create all the XML files ourselves
from the test package, and then zip them up to a .docx file for
subsequent opening by what will then be the only public Word APIs.

The only test case to be changed is images/word/update-drawingml01.test,
which was necessarily because it assumed a particular order of
relationships in document.xml.rels, which is changed now because we
already have styles.xml as a relationship *before* the image is added in
the update. The change simply reorders these, but the file is otherwise
semantically equivalent.