Word: fromPlain: Method to avoid WordPackage
Add a new implementation of Word_fromPackage (which, as before, takes a
TextPackage as input), this time creating the files directly instead of
using the WordPackage class. We still use the OPC APIs however for
saving relationships asnd content types; this is a transitional step.

This commit includes both the old and new versions of Word_fromPackage,
and runs both of them, checking that the output of each is identical.
The old version will be removed once the new version has been modified
to write relationship files directly instead of via the OPC APIs. All
tests pass, including the output comparisons of the two functions.