Word: toPlain: Don't use WordPackage/OPC in impl
When converting a word document to plain text, avoid using any of the
WordPackage or OPC APIs, with the exception of saving the package to a
temporary file in the process. This paves the way for us completely
avoiding these APIs (which whould be internal to DocFormats) within
dfutil.

Because Word_toPlain must first save the package and unzip it somewhere,
it now requires the caller to supply a temporary directory. This
eventually won't be necessary once we come up with an in-memory abstract
storage API which allows the code to work with filesystem-like data
without actually having to store it on the filesystem.