Add DFStore
The DFStore class provides a VFS-style abstraction over a set of files
and directories. Up until now, all of the conversion code has used
temporary directories to store files sourced or destined for the zipped
.docx format.  Because there's not yet an easy API for accessing
individual entries in a zip file (and zip files must be written in bulk,
not by individual entry), all the work has gone on in a temporary
directory.

The downside to this is that it forces unnecessary use of the
filesystem, which slows things down and introduces the risk of temporary
directories not being cleaned up if the program crashes. We ultimately
want to have all temporary data reside in memory only, which is both
faster and does not leave any data around on the filesystem if cleanup
does not occur due to a crash.

The DFStore class currently provides an abstraction over a physical
directory, but in the future will be expanded to also allow the storage
to be in-memory, or a zip file. When these capabilities are implemented,
it will be possible for the conversion code to store temporary files in
memory, and to read directly from the zip file without first having to
uncompress it to the filesystem.