DFStore: Use a vtable for multiple implementations
Treat DFStore as an abstract class, which can have multiple
implementations, by adding a vtable to the struct for each of the
operations.

Currently only the filesystem implementation exists (which is identical
to what was there before, but moved into separate functions). However
in-memory and zip file implementations will soon join it.