DFStore: Remove concept of directories
To simplify working with the store abstraction, as well as the upcoming
implementation of the in-memory store, expose the data as a flat set of
files, with path names matching a directory hierarchy. There is no
longer a concept of a directory; each file just has a path name that
identifies it. The filsystem store maps this to directories, but this
will not be necessary for other implementations, such as zip files and
in-memory stores.

In particular, this change means we will be able to implement the
in-memory store as a simple hash table mapping from filenames to buffer
objects. This avoids a complicated mechanism for representing a
directory tree.

As part of this change, writing to a file now creates its parent
directory if it does not already exist. Thus numerous DFStoreMkdir calls
have been removed, as has the function itself.

DFStoreList now returns only filenames, not directory names. It always
returns *every* file in the store, removing the 'recursive' and 'path'
options.