DFStore: In-memory implementation
Add a new implementation of DFStore that keeps all data in memory,
instead of a temporary directory. The implementation uses a simple hash
table mapping path names to DFBuffer objects.

This simplifies many parts of dfutil that had to first create a
temporary directory, ensure it was empty, and then delete it afterwards.
It also avoids the problem that due to the lack of mkdtemp on Windows,
we always name the directory "dftemp", which doesn't work if we want to
use multiple threads.