WordPackageOpen{New,From}: Return new object
Previously, we had separate methods for allocating a WordPackage object
and subsequently initialising it with an empty document or the contents
of an existing file. This required code like the following:

    WordPackage *package = WordPackageNew(store);
    if (!WordPackageOpenFrom(package,filename,error)) {
        // ... Report error
        WordPackageRelease(package);
    }
    else {
        // ... Do stuff with package
        WordPackageRelease(package);
    }

Now, we no longer expose WordPackageNew, and instead merge that with
WordPackageOpenNew and WordPackageOpenFrom, which now return the new
WordPackage object on success, or NULL on failure. So the above code
would now look like:

    WordPackage *package = WordPackageOpenFrom(store,filename,error);
    if (package == NULL) {
        // ... Report error
    }
    else {
        // ... Do stuff with package
        WordPackageRelease(package);
    }

Although it makes only a small difference with the current API (one
function call instead of two, easier cleanup), this change has mostly
been done in preparation for a generic open/new function which doesn't
expose an explicit document format to the caller.