OPCPackage: Do init in constructors
Change OPCPackageOpenNew and OPCPackageOpenFrom so they both create
*and* initialise the object. Previously, it was necessary to call
OPCPackageNew first, and then pass that object to one of the above
functions. Instead of returning an integer to indicate success or
failure, they now return the package object on success, or NULL on
failure.

This change mirrors the one made to WordPackage in the previous commit.