XML: Move int HTML_requiresCloseTag to DFXML.c
This was the only function from HTML which was needed in xml, and xml
was the only thing that used it. Thus is really belongs in XML. Moving
it here and making it internal to DFXML.c means that we can remove all
the CSS and HTML includes from this file.