Organise source tree into modules
Rearrange the source files into a directory hierarchy that groups them
clearly into different "modules". Each module should have a src,
headers, and tests directory; some of these directories are missing but
will be added in due course.

The modules we have in DocFormats are as follows:

    3rdparty         - External libraries developed by others
    platform         - Abstraction layer over platform-specific APIs
    core             - Fnfrastructure for use by modules and utilities
    filters/html     - Currently only contains test cases
    filters/latex    - LaTeX implementation (currently export only)
    filters/ooxml    - OOXML implementation (currently .docx only)
    filters/odf      - ODF implementation (currently very early stages)

The modules we have outside of DocFormats are as follows:

    consumers/dfutil - General purpose test + conversion program

This gives us the following directory structure:

    Corinthia
     |- consumers
     |   \- dfutil
     \- DocFormats
         |- 3rdparty
         |- core
         \- filters
             |- html
             |- latex
             |- odf
             \- ooxml

Other parts still remain to be added, including missing directories in
some modules, plus an api module for interfacing with consumers of the
library. This commit just moves existing files and adjusts the build
files appropriately.