DFInitOnce: Avoid dependency on pthread/windows.h
Previously, we needed to include pthread.h so that the pthread_once_t
type would be defined, which the compiler would use to determine how
much storage to allocate for a one-time initialisation control variable
(e.g. as used in DFNameMap.c). However, we want to avoid having to
include this header file directly or indirectly in any code outside the
platform module.

To get around this requirement, we now implement a custom one-time
execution function in DFInitOnce. We define the DFOnce type (previously
corresponding to pthread_once_t) as int, and DF_ONCE_INIT (previously
PTHREAD_ONCE_INIT) as 0. In Unix.c (as well as Win32.c) we use a mutex
to control access to the int variable, so we can ensure the test and set
happens atomically, allowing us to guarantee that the function executes
only once.