Move DFErrorSetWin32 to Win32.c
This function was previously prototyped in DFError.h, which required
inclusion of <windows.h> (via DFTypes.h) in order for the DWORD type to
be recognised. To avoid having to include <windows.h> indirectly in all
header files, and because this function is only actually used in
Win32.c, move it there instead.