Move system includes out of DFCommon.h
Instead, only include system headers in C source files.

We still have a few bits and pieces in DFCommon.h to paper over platform
differences. For windows, we need to define snprintf and strasecmp to
their windows equivalents (_snprintf and _stricmp). And the S_ISDIR
macro is missing on Windows so that is still there, but we can probably
use the replacement instead since it's only needed in a couple of
places.

Ideally we want to avoid all #ifdef WIN32 everywhere, and abstract that
in the platform layer. Currently however we need to include <direct.h>
(windows) and <unistd.h> (Unix) in a couple of places since certain
functions like rmdir are defined in different headers in the two
platforms. I'm not sure what the best way to deal with this is, in the
absence of including those in DFCommon.h, but this is only in a handful
of places within the code right now.