Rename DFStore to DFPackage
'Package' is a more suitable term for this class, as is refers to the
idea of a collection of files stored in some packaging format (which
could be a zip file, directory, or memory). It also corresponds to the
terminology used in both the ODF and OOXML specs.

In cases where we use a class like WordPackage, which operates at a
higher level of abstraction, in conjunction with a DFPackage object, we
refer to the latter as a "raw" package, in the sense that it just
provides storage capabilities without any semantic functionality.