Word_to/fromPlain: Add rawPackage parameter
When converting from a plain text package to a WordPackage, also pass
back the raw DFPackage object, so that the caller can later use it to
get at the resulting package contents after modification.

When converting from a WordPackage to a plain text package, do so using
the DFPackage object.