WordPackage: Don't explicitly zip/unzip file
In WordPackage and OPCPackage, assume that the contents of the DFPackage
object can be loaded and save directly, without the need to first unzip
or zip the data from a file. This allows us to work with packages
resresented solely in memory, e.g. in the test cases. If the data is to
actually be stored in a physical .docx file (which is just a zip file),
then the code that uses WordPackage must manage this.

This is a temporary measure until we have a DFPackage implementation
that can work directly with zip files.