Add zip-based DFPackage
This new "subclass" of DFPackage operates with a zip file, allowing code
that loads and stores data from DFPackage objects to work directly with
zip files, rather there having to be an explicit DFUnzip call beforehand
or a DFZip call afterwards. As such, code like the word converter
doesn't need to actually care whether the package contents originally
came from or a destined for a zip file; it just deals with packages in
the abstract sense.

Currently, the implementation actually operates just like a memory
package, unzipping the contents when created, and creating a new zip
file on save. It is envisaged however that eventually this will be
changed to read from an existing zip file on-demand.