Replace DFPackageNewZip with Create/Open functions
Distinguish between the act of creating a new zip file and opening an
existing one (for reading and/or modification).

The separation between these two functions takes the place of the
mustExist parameter that was accepted by DFPackageNewZip, and also makes
the creation case clearer as it returns an error if the file already
exists. The latter thus prevents accidentally bringing in one or more
files from an old zip file if the intention is to create an entirely new
one.