Add DFFileFormat enum
This defines current and future supported file formats, and comes with
functions to convert to and from a file extension.

Each DFPackage now has a file format associated with it. For those
mapped to a zip file, the format is determined by the file extension.
For those created in memory or mapped to a directory, this must be
specified by the caller at creation time.

The purpose of recording the format of a package is so that a
determination can be made whether it is valid to use in a given
situation (e.g. if it is to be treated as a word document), or what
action to take based on the type (e.g. a different code path must be
taken for docx files vs. odt files).