API classes for concrete & abstract documents
On top of the simple API that specifies get, put, and create functions
based solely on filenames, we now have corresponding functions which
work on concrete and abstract document objects.

These objects are wrappers around a DFPackage, which in turn is a
wrapper around a zip file, directory, or in-memory hash table. They are
intended for where more "custom" conversion situations arise, such as
when you have the contents of a docx file extracted to a directory
(which is what UX Write currently does), or for the test cases where we
construct an in-memory DFPackage from a text-based representation of a
docx file.

Currently the DFGet, DFPut, and DFCreate functions are not yet
implemented. When they are, the DFGetFile, DFPutFile, and DFCreateFile
functions will become wrappers around them.