Avoid referencing WordPackage in dfutil
The WordPackage class should be private to the DocFormats library, and
accessed only via the public APIs. There were numerous places in the
test code where this was not the case. With this commit, we now have a
cleaner separation between the two, using the recently-added DFGet,
DFPut, and DFCreate APIs for performing bidirectional transformation on
documents.

The test functions in particular illustrate the utility of the DFPackage
abstraction, in that they create packages in memory only, without having
to write out to the filesystem. This simplifies the test code by in most
cases removing the need for temporary directories.

One place where we do sort-of break this abstraction is the tests for
expanding and collapsing bookmarks in a Word document. This isn't
something that should be part of the public API, but nonetheless needs
to be covered by the tests. Although we do bypass the API for this, we
use the new functions WordExpandBookmarks and WordCollapseBookmarks
declared in Word.h for this purpose, which take a DFPackage object, not
a WordPackage object parameters. It's likely there will be many other
similar cases like this in the future, and we still need to figure out a
way to address testing of purely internal functionality.