Move CSSClassNames.c into Word filter
This file references the WordSetupTableGridStyle function defined in the
OOXML filter, which we're not supposed to be accessing from core.
Although all the rest of the code here is generic in nature, in that it
could conceivably be used by other filters in the future, to resolve the
dependency issue we'll put this file into the OOXML filter for the time
being.