Move text package functions into DocFormats
The Word and HTML test code that currently lives in dfutil (but not for
much longer) uses plain text representations of "packages", which are
collections of files e.g. like you might find in a .docx file or a
directory with a HTML file and some images. With the new directory
organisation where we keep tests in the DocFormats tree itself (so as to
not require access to public APIs from an external program), we need
this functionality in the library before we can move the test functions
in.

Currently, the actual test functions which use these plain text
representations are still in dfutil; however they will soon be moved to
the tests directory within the the html and ooxml directories.