Use dotted notation for function names in *.test
In the *.test data files, the first line of the file (after comments)
indicates the test function to execute. Instead of the old, haphazard
naming convention, we now use the form module.testfunction, where the
module can contain further dots, e.g. ooxml.word.get.

This was already in place from the previous commit but the *.test files
still referenced the old names, necessitating the use of a substitution
mechanism. The files all now use the new naming scheme, and the
substitution mechanism has been removed.