.gitignore excluding Microsoft ephemeral artifacts
The .gitignore is expanded to include exclusions that GitHub provides by
default when creating a git repository used for Windows development.
This annotated version includes some miscellaneous catches, such as
temporary files produced by a variety of editors.   In addition,
exclusion of Visual Studio Solution and Project files is added, with
explanatory comment.