Remove NamesTests.c to fix compilation on windows
The names directory contains only constants for representing particular
element names, and doesn't need tests. Removing NamesTests.c for the
time being as it inadvertently re-defined the LibTests symbol, which
caused compilation to fail on windows.

The way we use integer values to identify elements in the DOM tree needs
to be re-thought anyway, as it doesn't scale or separate for different
file formats. For example there's no situation in which you would have
ODF and OOXML elements in the same XML document; there needs to be a
separate "namespace", to to speak, of integer tags for each filter, with
common ones like HTML included.