Don't include <io.h> on non-Windows platforms
The <io.h> system header is only avilable on windows. It was
unconditionally included in DFFilesystem.c, which broke the build on
Linux and OS X. It used to be wrapped in an #ifdef WIN32 but with the
removal of this, it was already included.

I have now moved the #include <io.h> to DFPlatform.h, where there is an
ifdef in place to distinguish between WIN32 and other platforms, so that
we don't get a compile error when this is built on Linux and OS X.