Fix OS X build breakage in DFError.h
Commit e8d93d4 removed the ATTRIBUTE_FORMAT macro from DFError.h, since
this macro is also present in DFPlatform.h. However, DFError.h is part
of the public API, while DFPlatform is not. Consumers that are
configured to compile only against the public API are unable to access
macro definitions in DFPlatform.h, so those macros must be present in
the API headers.

The ATTRIBUTE_FORMAT is used for telling the compiler to check the
arguments to DFErrorFormat to make sure they conform to printf-style
formatting. GCC/Clang and Visual C++ both have a different syntax for
this, which is why a macro is needed.

I'm not sure why gcc on Linux and Visual C++ on windows weren't
complaining about this, but Xcode gave a compile error as the macro was
undefined.

If we want to avoid platform-specific #ifdefs in specific public header
files, we'll need something equivalent to DFPlatform.h which is included
in the public headers. Putting this #ifdef back as-is is a temporary
solution, needed to build the code under Xcode, but ultimately it would
be good if we could come up with a more general solution.