Take ATTRIBUTE_FORMAT macros back out of DFError.h
This macro is in DFPlatform.h, which all source files include. However
there was a build failure on OS X, which turned out to be the two
Objective C files (FunctionTests.m and StringTests.m) which did not have
the DFPlatform.h include.

Both files now have this include, so the macro does not need to repeated
in DFError.h.