Update copyright notice
Place the copyright notice described at:

    http://www.apache.org/legal/src-headers.html

in all source and header files, with the exception of those in the
platform directory (which will be taken care of in a pending merge by
Jan). This notice takes place of the previous one which specifically
mentioned UX Productivity.

Also add a NOTICE.txt file based on the format described in the above
link. The copyright ownership statement for UX Productivity now lives in
this file. Anyone else who has made contributions to the project should
add their own statement to this file.