Enable hit testing fixes for above/below body rect
At the start of Position_atPoint, we check if the position is above or
below the body rect, and if so, change the y value so it's on the first
or last line. Previously, we would then just go ahead and return the
result of document.caretRangeFromPoint, but this would skip the other
checks later in the function for things like images and empty footnotes
and endnotes. Now we just adjust the y position, and then continue on
with these checks.