Fix selection hit test after footnotes/endnotes
Hit testing in Position_atPoint is a bit more complicated when we have
selection spans in the tree. When looking for "special" nodes like empty
footnotes or endnotes, we have to take into account the possibility that
they might be contained within a selection span. Additionally, we also
have to consider that the result of document.caretRangeFromPoint might
also be inside a selection span, in which case the previous/next nodes
we want to look at are that of the span, not the node inside it.

The newly-added posOutsideSelection function checks a given position to
see if it's at the very beginning or end of a selection span. If so, it
returns a position that is directly before or after the span. Otherwise,
if the position is part-way through the span (e.g. within a text node),
the original position is returned.

We call this function with the result of document.caretRangeFromPoint,
but *only* for the purposes of checking the previous and next nodes to
see if they are images or empty footnotes/endnotes. We do *not* use it
for other purposes, as doing so results in an odd flickering effect
where consecutive touch movements result in the selection end boundary
alternately jumping before and after the character that is touched.